replace the crank encoder with msg
This replaces @t0mg's original crank encoder with msg, new in Playdate OS 2.4+.  We're using @cwalther's concept for encoding data over msg, supporting printable characters as-is, numbers 32-255 inclusive to replicate the crank encoder model, and base64-encoded anything else, which will be useful for control characters.  The base64 library is @iskolbin's, and it needed @DimitriBarronmore's pd-require to load.